+++ sdk/identity/Azure.Identity/tests/samples/ReadmeSnippets.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT License.
 
 using System;
+using Azure.Core;
 using Azure.Messaging.EventHubs.Producer;
 using Azure.Security.KeyVault.Secrets;
 using Azure.Storage.Blobs;
@@ -36,19 +37,42 @@ namespace Azure.Identity.Samples
         }
 
         [Test]
-        public void UserAssignedManagedIdentity()
+        public void UserAssignedManagedIdentityWithClientId()
         {
             string userAssignedClientId = "";
 
-            #region Snippet:UserAssignedManagedIdentity
+            #region Snippet:UserAssignedManagedIdentityWithClientId
 
-            // When deployed to an azure host, the default azure credential will authenticate the specified user assigned managed identity.
+            // When deployed to an Azure host, DefaultAzureCredential will authenticate the specified user-assigned managed identity.
 
-            //@@string userAssignedClientId = "<your managed identity client Id>";
-            var credential = new DefaultAzureCredential(new DefaultAzureCredentialOptions { ManagedIdentityClientId = userAssignedClientId });
+            //@@string userAssignedClientId = "<your managed identity client ID>";
+            var credential = new DefaultAzureCredential(
+                new DefaultAzureCredentialOptions
+                {
+                    ManagedIdentityClientId = userAssignedClientId
+                });
 
-            var blobClient = new BlobClient(new Uri("https://myaccount.blob.core.windows.net/mycontainer/myblob"), credential);
+            var blobClient = new BlobClient(
+                new Uri("https://myaccount.blob.core.windows.net/mycontainer/myblob"),
+                credential);
 
+            #endregion
+        }
+
+        [Test]
+        public void UserAssignedManagedIdentityWithResourceId()
+        {
+            #region Snippet:UserAssignedManagedIdentityWithResourceId
+            string userAssignedResourceId = "<your managed identity resource ID>";
+            var credential = new DefaultAzureCredential(
+                new DefaultAzureCredentialOptions
+                {
+                    ManagedIdentityResourceId = new ResourceIdentifier(userAssignedResourceId)
+                });
+
+            var blobClient = new BlobClient(
+                new Uri("https://myaccount.blob.core.windows.net/mycontainer/myblob"),
+                credential);
             #endregion
         }