+++ sdk/openai/Azure.AI.OpenAI/src/Custom/ChatCompletions/ChatMessageImageContentItem.cs
@@ -2,29 +2,62 @@
 // Licensed under the MIT License.
 
 using System;
+using System.IO;
 
 namespace Azure.AI.OpenAI;
 
 public partial class ChatMessageImageContentItem : ChatMessageContentItem
 {
     // CUSTOM CODE NOTE:
-    //   This addition allows the direct use of a Uri with the content item constructor rather than needing an
-    //   intervening instantiation of the Image URL type.
+    //   - Hide the ImageUrl property for easier, direct use of the content item
+    //   - Provide custom Uri, BinaryData, and Stream constructors
 
-    /// <summary> Initializes a new instance of ChatMessageImageContentItem using default detail. </summary>
-    /// <param name="imageUri"> An internet location, which must be accessible to the model,from which the image may be retrieved. </param>
-    /// <exception cref="ArgumentNullException"> <paramref name="imageUri"/> is null. </exception>
-    public ChatMessageImageContentItem(Uri imageUri)
-        : this(new ChatMessageImageUrl(imageUri))
-    { }
+    internal ChatMessageImageUrl ImageUrl { get; }
 
-    /// <summary> Initializes a new instance of ChatMessageImageContentItem. </summary>
-    /// <param name="imageUri"> An internet location, which must be accessible to the model,from which the image may be retrieved. </param>
+    /// <summary>
+    /// Initializes a new instance of ChatMessageImageContentItem that refers to an image at another location via URL.
+    /// </summary>
+    /// <remarks>
+    /// This constructor should only be used for file references. To use binary data, streams, or a file directly,
+    /// please refer to the alternate constructors.
+    /// </remarks>
+    /// <param name="imageUri"> An internet location, which must be accessible to the model, from which the image may be retrieved. </param>
     /// <param name="detailLevel"> The image detail level the model should use when evaluating the image. </param>
     /// <exception cref="ArgumentNullException"> <paramref name="imageUri"/> is null. </exception>
-    public ChatMessageImageContentItem(Uri imageUri, ChatMessageImageDetailLevel detailLevel)
-        : this(imageUri)
+    public ChatMessageImageContentItem(Uri imageUri, ChatMessageImageDetailLevel? detailLevel = null)
+        : this(new ChatMessageImageUrl(imageUri, detailLevel))
+    {}
+
+    /// <summary>
+    /// Initializes a new instance of ChatMessageImageContentItem from a BinaryData instance containing image
+    /// information in a known format.
+    /// </summary>
+    /// <param name="bytes"> The image data to provide as content. </param>
+    /// <param name="mimeType"> The MIME type, e.g. <c>image/png</c>, matching the format of the image data. </param>
+    /// <param name="detailLevel"> The image detail level the model should use when evaluating the image. </param>
+    public ChatMessageImageContentItem(BinaryData bytes, string mimeType, ChatMessageImageDetailLevel? detailLevel = null)
+        : this(new ChatMessageImageUrl(bytes, mimeType, detailLevel))
+    {}
+
+    /// <summary>
+    /// Initializes a new instance of ChatMessageImageContentItem from a BinaryData instance containing image
+    /// information in a known format.
+    /// </summary>
+    /// <param name="stream"> The image data to provide as content. </param>
+    /// <param name="mimeType"> The MIME type, e.g. <c>image/png</c>, matching the format of the image data. </param>
+    /// <param name="detailLevel"> The image detail level the model should use when evaluating the image. </param>
+    public ChatMessageImageContentItem(Stream stream, string mimeType, ChatMessageImageDetailLevel? detailLevel = null)
+        : this(new ChatMessageImageUrl(stream, mimeType, detailLevel))
+    {}
+
+    /// <summary> Initializes a new instance of <see cref="ChatMessageImageContentItem"/>. </summary>
+    /// <param name="imageUrl"> An internet location, which must be accessible to the model,from which the image may be retrieved. </param>
+    /// <exception cref="ArgumentNullException"> <paramref name="imageUrl"/> is null. </exception>
+    internal ChatMessageImageContentItem(ChatMessageImageUrl imageUrl)
     {
-        ImageUrl.Detail = detailLevel;
+        Argument.AssertNotNull(imageUrl, nameof(imageUrl));
+
+        Type = "image_url";
+        ImageUrl = imageUrl;
     }
 }