+++ sdk/openai/Azure.AI.OpenAI/tests/Samples/ChatWithVision.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT License.
 
 using System;
+using System.IO;
 using System.Threading.Tasks;
 using Azure.Identity;
 using NUnit.Framework;
@@ -14,20 +15,23 @@ public partial class ChatWithVision
     [Ignore("Only verifying that the sample builds")]
     public async Task ChatWithVisionPreview()
     {
-        string endpoint = "https://myaccount.openai.azure.com/";
+        string endpoint = "https://myresource.openai.azure.com/";
         var client = new OpenAIClient(new Uri(endpoint), new DefaultAzureCredential());
 
         #region Snippet:AddImageToChat
         const string rawImageUri = "<URI to your image>";
+        using Stream jpegImageStream = File.OpenRead("<path to a local image file>");
+
         ChatCompletionsOptions chatCompletionsOptions = new()
         {
-            DeploymentName = "gpt-4-vision-preview",
+            DeploymentName = "gpt-4-turbo",
             Messages =
             {
                 new ChatRequestSystemMessage("You are a helpful assistant that describes images."),
                 new ChatRequestUserMessage(
-                    new ChatMessageTextContentItem("Hi! Please describe this image"),
-                    new ChatMessageImageContentItem(new Uri(rawImageUri))),
+                    new ChatMessageTextContentItem("Hi! Please describe these images"),
+                    new ChatMessageImageContentItem(new Uri(rawImageUri)),
+                    new ChatMessageImageContentItem(jpegImageStream, "image/jpg", ChatMessageImageDetailLevel.Low)),
             },
         };
         #endregion
@@ -35,7 +39,7 @@ public partial class ChatWithVision
         #region Snippet:GetResponseFromImages
         Response<ChatCompletions> chatResponse = await client.GetChatCompletionsAsync(chatCompletionsOptions);
         ChatChoice choice = chatResponse.Value.Choices[0];
-        if (choice.FinishDetails is StopFinishDetails stopDetails || choice.FinishReason == CompletionsFinishReason.Stopped)
+        if (choice.FinishReason == CompletionsFinishReason.Stopped)
         {
             Console.WriteLine($"{choice.Message.Role}: {choice.Message.Content}");
         }