+++ sdk/search/Azure.Search.Documents/tests/Models/ReflectableStructModel.cs
@@ -57,6 +57,8 @@ namespace Azure.Search.Documents.Tests
 
         public sbyte SByte { get; set; }
 
+        public byte Byte { get; set; }
+
         public short Short { get; set; }
 
         public bool Flag { get; set; }
@@ -122,6 +124,16 @@ namespace Azure.Search.Documents.Tests
         public IEnumerable<sbyte> SByteIEnumerable { get; set; }
 
         public ICollection<sbyte> SByteICollection { get; set; }
+
+        public byte[] ByteArray { get; set; }
+
+        public IList<byte> ByteIList { get; set; }
+
+        public List<byte> ByteList { get; set; }
+
+        public IEnumerable<byte> ByteIEnumerable { get; set; }
+
+        public ICollection<byte> ByteICollection { get; set; }
 
         public short[] ShortArray { get; set; }